feat: convert app router files to TypeScript
- app/layout.tsx: typed children as ReactNode
- app/page.tsx, projects/page.tsx, articles/.../page.tsx: server components with typed metadata
- Client components preserved "use client" directive
