--- app/articles/how-to-set-up-spf-dkim-dmarc/_ArticleClient.js
@@ -1,8 +1,0 @@
-"use client"
-
-import Article from "@/layouts/Article"
-import ArticleContent from "../../../content/articles/how-to-set-up-spf-dkim-dmarc.mdx"
-
-export default function ArticleClient() {
-  return <Article><ArticleContent /></Article>
-}
