+++ app/articles/how-to-set-up-spf-dkim-dmarc/_ArticleClient.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import Article from '@/layouts/Article'
+import ArticleContent from '../../../content/articles/how-to-set-up-spf-dkim-dmarc.mdx'
+
+export default function ArticleClient() {
+  return <Article><ArticleContent /></Article>
+}
