+++ app/articles/how-to-set-up-spf-dkim-dmarc/page.tsx
@@ -1,8 +1,8 @@
-import { createMetadata } from "@/lib/metadata"
-import ArticleClient from "./_ArticleClient"
+import { createMetadata } from '@/lib/metadata'
+import ArticleClient from './_ArticleClient'
 
 export const metadata = createMetadata(
-  "I figured out how DMARC works, and it almost broke me",
+  'I figured out how DMARC works, and it almost broke me',
   "How to use SPF, DKIM, and DMARC to improve your domain's email security and limit spoofing - but written for humans."
 )
