--- app/layout.js
@@ -1,9 +1,0 @@
-import "../theme/index.css"
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="en-GB">
-      <body>{children}</body>
-    </html>
-  )
-}
