+++ app/layout.tsx
@@ -1,0 +1,11 @@
+import type { ReactNode } from 'react'
+
+import '../theme/index.css'
+
+export default function RootLayout({ children }: { children: ReactNode }) {
+  return (
+    <html lang="en-GB">
+      <body>{children}</body>
+    </html>
+  )
+}
