--- app/projects/_ProjectsClient.js
@@ -1,8 +1,0 @@
-"use client"
-
-import StandardLayout from "@/layouts/StandardLayout"
-import ProjectsContent from "../../content/projects.mdx"
-
-export default function ProjectsClient() {
-  return <StandardLayout><ProjectsContent /></StandardLayout>
-}
