+++ app/projects/_ProjectsClient.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import StandardLayout from '@/layouts/StandardLayout'
+import ProjectsContent from '../../content/projects.mdx'
+
+export default function ProjectsClient() {
+  return <StandardLayout><ProjectsContent /></StandardLayout>
+}
